Separate TabRegistry from its factory to prevent unneeded dependencies in unit tests
